DDD/OLD_PROJECTS/requirements.txt
# AI Valuation Model Requirements
# Python dependencies for Bulgarian car valuation AI model

# Google Cloud
google-cloud-bigquery>=3.11.0
google-cloud-aiplatform>=1.25.0
google-cloud-storage>=2.7.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=1.7.0
pandas>=1.5.0
numpy>=1.24.0

# Data Processing
joblib>=1.3.0

# Jupyter (for development)
jupyter>=1.0.0
ipykernel>=6.0.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0